feat: generate gen_ai_types.py from OpenTelemetry JSON schemas
