+++ python/mirascope/ops/_internal/instrumentation/llm/gen_ai_types/gen_ai_system_instructions.py
@@ -1,0 +1,18 @@
+"""OpenTelemetry Gen AI Semantic Conventions types."""
+
+from __future__ import annotations
+
+from typing import TypeAlias
+
+from . import shared
+
+SystemInstructions: TypeAlias = list[
+    shared.TextPart
+    | shared.ToolCallRequestPart
+    | shared.ToolCallResponsePart
+    | shared.BlobPart
+    | shared.FilePart
+    | shared.UriPart
+    | shared.ReasoningPart
+    | shared.GenericPart
+]
